+++ packages/Statie/packages/Generator/tests/GeneratorTest.php
@@ -42,53 +42,4 @@ final class GeneratorTest extends AbstractGeneratorTest
         $firstPost = $posts[0];
         $this->assertInstanceOf(DateTimeInterface::class, $firstPost['date']);
     }
-
-    public function testPostExceptionsOnUnset(): void
-    {
-        $post = $this->getPost();
-        $this->expectException(UnsupportedMethodException::class);
-        unset($post['key']);
-    }
-
-    public function testPostExceptionOnSet(): void
-    {
-        $post = $this->getPost();
-        $this->expectException(UnsupportedMethodException::class);
-        $post['key'] = 'value';
-    }
-
-    public function testPostExceptionOnGetNonExistingSuggestion(): void
-    {
-        $post = $this->getPost();
-
-        $this->expectException(AccessKeyNotAvailableException::class);
-        $this->expectExceptionMessage(sprintf(
-            'Value "tite" was not found for "%s" object. Did you mean "title"?',
-            PostFile::class
-        ));
-
-        $value = $post['tite'];
-    }
-
-    public function testPostExceptionOnGetNonExistingAllKeys(): void
-    {
-        $post = $this->getPost();
-
-        $this->expectException(AccessKeyNotAvailableException::class);
-        $this->expectExceptionMessage(sprintf(
-            'Value "key" was not found for "%s" object. Available keys are: "id", "title", "relativeUrl".',
-            PostFile::class
-        ));
-
-        $value = $post['key'];
-    }
-
-    private function getPost(int $order = 0): PostFile
-    {
-        $this->generator->run();
-
-        $posts = $this->configuration->getOption('posts');
-
-        return $posts[$order];
-    }
 }
